+++ apps/renderer/src/modules/rsshub/add-modal-content.tsx
@@ -1,0 +1,98 @@
+import { Button } from "@follow/components/ui/button/index.js"
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@follow/components/ui/form/index.jsx"
+import { Input } from "@follow/components/ui/input/Input.js"
+import { zodResolver } from "@hookform/resolvers/zod"
+import { useEffect } from "react"
+import { useForm } from "react-hook-form"
+import { useTranslation } from "react-i18next"
+import { z } from "zod"
+
+import { whoami } from "~/atoms/user"
+import { ShikiHighLighter } from "~/components/ui/code-highlighter"
+import { useShikiDefaultTheme } from "~/components/ui/code-highlighter/shiki/hooks"
+import { useAddRSSHubMutation } from "~/queries/rsshub"
+
+export function AddModalContent({ dismiss }: { dismiss: () => void }) {
+  const { t } = useTranslation("settings")
+  const addRSSHubMutation = useAddRSSHubMutation()
+  const shikiTheme = useShikiDefaultTheme()
+  const me = whoami()
+
+  const formSchema = z.object({
+    baseUrl: z.string().url(),
+    accessKey: z.string().optional(),
+  })
+
+  const form = useForm<z.infer<typeof formSchema>>({
+    resolver: zodResolver(formSchema),
+  })
+
+  const onSubmit = (data: z.infer<typeof formSchema>) => {
+    addRSSHubMutation.mutate(data)
+  }
+
+  useEffect(() => {
+    if (addRSSHubMutation.isSuccess) {
+      dismiss()
+    }
+  }, [addRSSHubMutation.isSuccess])
+
+  return (
+    <div className="max-w-[550px] space-y-4 lg:min-w-[550px]">
+      <div className="text-sm">{t("rsshub.addModal.description")}</div>
+
+      <ShikiHighLighter
+        transparent
+        theme={shikiTheme}
+        className="group relative mt-3 cursor-auto select-text whitespace-pre break-words rounded-lg border border-border bg-zinc-100 p-2 text-sm dark:bg-neutral-800 [&_pre]:whitespace-pre [&_pre]:break-words [&_pre]:!p-0"
+        code={`FOLLOW_OWNER_USER_ID=${me?.handle || me?.id}               # User id or handle of your follow account
+FOLLOW_DESCRIPTION=${me?.name}'s instance # The description of your instance
+FOLLOW_PRICE=0                            # The monthly price of your instance, set to 0 means free.
+FOLLOW_USER_LIMIT=1000                    # The user limit of your instance, set it to 0 or 1 can make your instance private, leaving it empty means no restriction`}
+        language="dotenv"
+      />
+      <Form {...form}>
+        <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
+          <FormField
+            control={form.control}
+            name="baseUrl"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>{t("rsshub.addModal.base_url_label")}</FormLabel>
+                <FormControl>
+                  <Input placeholder="https://" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="accessKey"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>{t("rsshub.addModal.access_key_label")}</FormLabel>
+                <FormControl>
+                  <Input {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <div className="flex items-center justify-end">
+            <Button type="submit" isLoading={addRSSHubMutation.isPending}>
+              {t("rsshub.addModal.add")}
+            </Button>
+          </div>
+        </form>
+      </Form>
+    </div>
+  )
+}
