+++ apps/renderer/src/modules/rsshub/set-modal-content.tsx
@@ -1,0 +1,148 @@
+import { Button } from "@follow/components/ui/button/index.js"
+import { Card, CardContent } from "@follow/components/ui/card/index.js"
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@follow/components/ui/form/index.jsx"
+import { Input } from "@follow/components/ui/input/Input.js"
+import type { RSSHubModel } from "@follow/models"
+import { zodResolver } from "@hookform/resolvers/zod"
+import { useEffect } from "react"
+import { useForm } from "react-hook-form"
+import { useTranslation } from "react-i18next"
+import { z } from "zod"
+
+import { whoami } from "~/atoms/user"
+import { useAuthQuery } from "~/hooks/common"
+import { UserAvatar } from "~/modules/user/UserAvatar"
+import { Queries } from "~/queries"
+import { useSetRSSHubMutation } from "~/queries/rsshub"
+
+export function SetModalContent({
+  dismiss,
+  instance,
+}: {
+  dismiss: () => void
+  instance: RSSHubModel
+}) {
+  const { t } = useTranslation("settings")
+  const setRSSHubMutation = useSetRSSHubMutation()
+  const details = useAuthQuery(Queries.rsshub.get({ id: instance.id }))
+  const hasPurchase = !!details.data?.purchase
+  const price = instance.ownerUserId === whoami()?.id ? 0 : instance.price
+
+  const formSchema = z.object({
+    months: z.coerce
+      .number()
+      .min(hasPurchase ? 0 : 1)
+      .max(12),
+  })
+
+  const form = useForm<z.infer<typeof formSchema>>({
+    resolver: zodResolver(formSchema),
+    defaultValues: {
+      months: hasPurchase ? 0 : 1,
+    },
+  })
+
+  const months = form.watch("months")
+
+  const onSubmit = (data: z.infer<typeof formSchema>) => {
+    setRSSHubMutation.mutate({ id: instance.id, durationInMonths: data.months })
+  }
+
+  useEffect(() => {
+    if (setRSSHubMutation.isSuccess) {
+      dismiss()
+    }
+  }, [setRSSHubMutation.isSuccess])
+
+  return (
+    <div className="max-w-[550px] space-y-4 lg:min-w-[550px]">
+      <Card>
+        <CardContent className="max-w-2xl space-y-2 p-6">
+          <div className="mb-3 text-lg font-medium">{t("rsshub.useModal.about")}</div>
+          <div className="flex items-center gap-4">
+            <div className="text-sm text-muted-foreground">{t("rsshub.table.owner")}</div>
+            <UserAvatar
+              userId={instance.ownerUserId}
+              className="h-auto justify-start p-0"
+              avatarClassName="size-6"
+            />
+          </div>
+          <div className="flex items-center gap-4">
+            <div className="text-sm text-muted-foreground">{t("rsshub.table.description")}</div>
+            <div className="line-clamp-2">{instance.description}</div>
+          </div>
+          <div className="flex items-center gap-4">
+            <div className="text-sm text-muted-foreground">{t("rsshub.table.price")}</div>
+            <div className="line-clamp-2 flex items-center gap-1">
+              {instance.price} <i className="i-mgc-power text-accent" />
+            </div>
+          </div>
+          <div className="flex items-center gap-4">
+            <div className="text-sm text-muted-foreground">{t("rsshub.table.userCount")}</div>
+            <div className="line-clamp-2">{instance.userCount}</div>
+          </div>
+          <div className="flex items-center gap-4">
+            <div className="text-sm text-muted-foreground">{t("rsshub.table.userLimit")}</div>
+            <div className="line-clamp-2">{instance.userLimit || t("rsshub.table.unlimited")}</div>
+          </div>
+        </CardContent>
+      </Card>
+      {details.data?.purchase && (
+        <div>
+          <div className="text-sm text-muted-foreground">
+            {t("rsshub.useModal.purchase_expires_at")}
+          </div>
+          <div className="line-clamp-2">
+            {new Date(details.data.purchase.expiresAt).toLocaleString()}
+          </div>
+        </div>
+      )}
+      <Form {...form}>
+        <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
+          {price > 0 && (
+            <FormField
+              control={form.control}
+              name="months"
+              render={({ field }) => (
+                <FormItem>
+                  <FormLabel>{t("rsshub.useModal.months_label")}</FormLabel>
+                  <FormControl>
+                    <div className="flex items-center gap-10">
+                      <div className="space-x-2">
+                        <Input className="w-24" type="number" max={12} {...field} />
+                        <span className="text-sm text-muted-foreground">
+                          {t("rsshub.useModal.month")}
+                        </span>
+                      </div>
+                      <div className="flex items-center gap-1">
+                        {price * field.value}
+                        <i className="i-mgc-power text-accent" />
+                      </div>
+                    </div>
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+          )}
+          <div className="flex items-center justify-end">
+            <Button type="submit" isLoading={setRSSHubMutation.isPending}>
+              {price
+                ? t("rsshub.useModal.useWith", {
+                    amount: price * months,
+                  })
+                : t("rsshub.table.use")}
+            </Button>
+          </div>
+        </form>
+      </Form>
+    </div>
+  )
+}
