+++ locales/settings/en.json
@@ -284,7 +284,13 @@
   "profile.title": "Profile Settings",
   "profile.updateSuccess": "Profile updated.",
   "profile.update_password_success": "Password updated.",
-  "rsshub.public_instances": "Instances",
+  "rsshub.addModal.access_key_label": "Access Key (Optional)",
+  "rsshub.addModal.add": "Add",
+  "rsshub.addModal.base_url_label": "Base URL",
+  "rsshub.addModal.description": "To use your own instance in Follow, you must add the following environment variables to it.",
+  "rsshub.add_new_instance": "Add New Instance",
+  "rsshub.description": "RSSHub is a community-driven open-source RSS network. Follow provides a built-in dedicated instance and uses it to support thousands of subscription contents, you can also achieve more stable content acquisition by using your own or third-party instances.",
+  "rsshub.public_instances": "Available Instances",
   "rsshub.table.description": "Description",
   "rsshub.table.edit": "Edit",
   "rsshub.table.inuse": "In Use",
